requirements.txt
streamlit==1.32.2
azure-core==1.30.1
azure-ai-textanalytics==5.3.0
